Adding login requirement for CRUD in posts and reeds, backend

+++ posts/views.py
@@ -1,9 +1,8 @@
+from django.contrib.auth.decorators import login_required
 from django.shortcuts import render, get_object_or_404, redirect, reverse
-from django.views import generic
 from .models import Post
 from .forms import PostForm
-
-## CRUD views for Post
+from comments.forms import CommentForm
 
 # READ: List
 def post_list(request):
@@ -12,12 +11,30 @@ def post_list(request):
         "posts": posts
     })
 
-# READ: Detail
+# READ: Detail with comments
 def post_detail(request, pk):
     post = get_object_or_404(Post, pk=pk)
-    return render(request, 'posts/post_detail.html', {'post': post})
+    comments = post.comments.all()
+
+    if request.method == "POST":
+        form = CommentForm(request.POST)
+        if form.is_valid():
+            comment = form.save(commit=False)
+            comment.author = request.user
+            comment.content_object = post
+            comment.save()
+            return redirect('post_detail', pk=pk)
+    else:
+        form = CommentForm()
+
+    return render(request, 'posts/post_detail.html', {
+        'post': post,
+        'comments': comments,
+        'form': form,
+    })
 
 # CREATE
+@login_required
 def post_create(request):
     form = PostForm(request.POST or None)
     if form.is_valid():
@@ -28,8 +45,12 @@ def post_create(request):
     return render(request, 'posts/post_form.html', {'form': form})
 
 # UPDATE
+@login_required
 def post_update(request, pk):
     post = get_object_or_404(Post, pk=pk)
+    if request.user != post.author:
+        return redirect('post_detail', pk=pk)
+
     form = PostForm(request.POST or None, instance=post)
     if form.is_valid():
         form.save()
@@ -37,8 +58,12 @@ def post_update(request, pk):
     return render(request, 'posts/post_form.html', {'form': form})
 
 # DELETE
+@login_required
 def post_delete(request, pk):
     post = get_object_or_404(Post, pk=pk)
+    if request.user != post.author:
+        return redirect('post_detail', pk=pk)
+
     if request.method == 'POST':
         post.delete()
         return redirect('post_list')
